:art: 产品信息编辑功能

+++ src/main/java/com/rymcu/forest/mapper/ProductMapper.java
@@ -38,5 +38,5 @@ public interface ProductMapper extends Mapper<Product> {
      * @param type
      * @return
      */
-    ProductDTO selectProductDTOById(@Param("idProduct") Integer idProduct, @Param("type") Integer type);
+    ProductDTO selectProductDTOById(@Param("idProduct") Long idProduct, @Param("type") Integer type);
 }
